main: implement the "wait connection" pattern
The buzzer is expected to blink slowly when trying to connect. This
pattern is the only one that can't be driven by the backend, as we are
not connected yet, so hardcode a pattern command once the application is
started. While doing so, avoid having an immediate blink: delay the
first blink for a few seconds to give the opportunity to the buzzer to
"immediately show that it connects".

Signed-off-by: Alexis Lothoré <alexis.lothore@gmail.com>